+++ lib/scout_apm/agent_context.rb
@@ -205,7 +205,7 @@ module ScoutApm
       if !@config.any_keys_found?
         logger.info("No configuration file loaded, and no configuration found in ENV. " +
                     "For assistance configuring Scout, visit " +
-                    "https://help.apm.scoutapp.com/#ruby-configuration-options")
+                    "https://docs.scoutapm.com/#ruby-configuration-options")
       end
     end
   end