+++ lib/scout_apm/tracked_request.rb
@@ -42,6 +42,9 @@ module ScoutApm
     # the name is determined from the name of the Controller or Job layer.
     attr_accessor :name_override
 
+    # A unique, but otherwise meaningless String to identify this request. UUID
+    attr_reader :transaction_id
+
     # When we see these layers, it means a real request is going through the
     # system. We toggle a flag to turn on some slightly more expensive
     # instrumentation (backtrace collection and the like) that would be too
@@ -64,6 +67,7 @@ module ScoutApm
       @mem_start = mem_usage
       @recorder = agent_context.recorder
       @real_request = false
+      @transaction_id = ScoutApm::Utils::TransactionId.new.to_s
       ignore_request! if @recorder.nil?
     end
 
@@ -208,8 +212,8 @@ module ScoutApm
     def stop_request
       @stopping = true
 
-      if recorder
-        recorder.record!(self)
+      if @recorder
+        @recorder.record!(self)
       end
     end
 
@@ -275,6 +279,8 @@ module ScoutApm
 
       @agent_context.transaction_time_consumed.add(unique_name, root_layer.total_call_time)
 
+      context.add(:transaction_id => transaction_id)
+
       # Make a constant, then call converters.dup.each so it isn't inline?
       converters = {
         :histograms => LayerConverters::Histograms,
@@ -287,6 +293,11 @@ module ScoutApm
 
         :slow_job => LayerConverters::SlowJobConverter,
         :slow_req => LayerConverters::SlowRequestConverter,
+
+        # This is now integrated into the slow_job and slow_req converters, so that
+        # we get the exact same set of traces either way. We can call it
+        # directly when we move away from the legacy trace styles.
+        # :traces => LayerConverters::TraceConverter,
       }
 
       walker = LayerConverters::DepthFirstWalker.new(self.root_layer)